Corrected many collision and rounding bugs. Collisions are properly take into account for velocity calculation. Blocks are now properly retrieved from their coordinates. Increased ambient light.

+++ src/blocks/BlockInfo.h
@@ -20,7 +20,7 @@ public:
 
 private:
 
-	unsigned short int i_value;
+	int i_value;
 };
 
 inline bool operator==(const BlockInfo& block1, const BlockInfo& block2)
